requirements.txt
PyYAML==3.12
requests==2.12.3
slumber==0.7.1
